Make CompileResult.CompileError report accurate positions
Previously, the range positions were from the position relative to the
start of the statement. After this commit the positions are relative to
the section start (with `toUnsplicedPosition`) producing an absolute
position in the markdown document.

The benefit of this change is that we leave message formatting to
clients instead of forcing the compiler position formatting.