Implement vork:crash modifier
This commit implements support for asserting that a program fails with a
runtime exception. This modifier does not exist in tut because tut:fail
works for both runtime and compiler-time errors. The vork encoding makes
it impossible to support both runtime and compile-time errors with the
same modifier because all code fences are compiled together as one
compilation unit.