Gracefully handle position conversion errors
Previously, mdoc failed with an exception when converting bogus
positions reported by the compiler. Now we fall back to `Position.None`
instead, which means the message is reported without a caret.

The best long-term solution is probably to reconsider the architecture
for how `fail` modifiers are compiled.  The design in
https://github.com/olafurpg/mdoc/issues/95#issuecomment-426993507 is
both simpler than the current  macro-based design and should also
produce consistent errors with the compiler.