Change mdoc:fail encoding to use proper compiler
Previously, we used a macro to get compile errors for mdoc:fail.
However, this resulted in wrong compile error messages in rare cases.
Now, we produce  a full compilation unit for each mdoc:fail section
and compile as a regular source file.

A benefit of this new approach is that we report pretty range position
error messages instead of offset positions.

This change required customizing position formatting so I inlined the
Scalafix utils we were using and removed the Scalafix dependency.