Use objects instead of classes.
With classes, we get the following warning when declaring sealed ADTs:

```
warning: The outer reference in this type test cannot be checked at run time.
```

Previously, we wrapped the code in `class Session { class App { ... } }`.
Now, we wrap the code in `object Session { object App { ... } }`
instead, which removes the warning. Interestingly, this change improved
error reporting for `:fail` in one test case where I was unhappy with
the old behavior.