Automatically restart compiler when classfiles are missing.
Previously, mdoc reused the same compiler instance for all compilations.
Now, we create a new compiler instance when the following situation
happens:

- compilation succeeds without any reported errors
- compilation produced no classfiles

I've consistently been able to reproduce this situation in the
https://github.com/spotify/scio repo when running mdoc on all the
markdown sources but I haven't been able to minimize the error to a
single source file. In Metals we restart the compiler frequently to
avoid cryptic errors like this. The compiler has a lot of state that
tends to go bad after several repeated compile requests.

Fixes #164.