Use custom parser for markdown processing.
Previously, mdoc used the flexmark library for processing markdown
files. Since flexmark is a proper markdown parser, it may not always
understand the format of the document that mdoc was processing, for
example:

* revealjs slides inside HTML files
* custom markdown extensions that are not supported by commonmark

This commit introduces a new custom parser that only understands the
parts of markdown files that mdoc touches: code fences. Since mdoc
doesn't need to understand all markdown flavors to process code fences,
we can get away with a much simpler parser than flexmark (that
understands all of markdown).

With this change, it should be possible to use mdoc on any document as
long as the code fences start at the beginning of a line with the
triple-backtick "scala mdoc" syntax
````
```scala mdoc
println(42)
```
````

Everything outside the code fences is left unchanged,
character-by-character, so the underlying markup language is
preserved.