Introduce new `mdoc:nest` modifier for redefining variables. (#251)
* Introduce new `mdoc:nest` modifier for redefining variables.

See https://underscore.io/blog/posts/2019/12/12/tut-to-mdoc.html for a
motivation.

Previously, the best workaround for redefining the same variable name
was to use `reset`, which was not a great workaround because it clears
all the existing statements in scope. Now, uses can use the `nest`
modifier to introduce a new scope that allows redefining the same
variable names while still accessing existing variables in scope.