Store compilers with dependencies *and* scalacOptions.
The main reason behind this change is that in Metals when a user adds or
removes a `$scalac`  magic import, it has no affect unless they close
Metals and restart it. This leads to people assuming that the `$scalac`
imports aren't working in Metals.

This is because when we do a looking into the ctx compilers we are only
taking into consideration the dependencies instead of the scalacOptions.
This slight changes just makes the key `(Dependencies, scalacOptions)`
instead of `Dependencies` so that when a user changes the `$scalac`
import, it's taken into consideration when doing the lookup or providing
a new compiler.