Update scalameta to 4.4.4
While updating it I also removed some deprecated usages of `Dotty` dialect and the dialect.copy method. Default toplevelSeparator
is "" so I removed that too.

There we also some issues with the newely added lineCaret and lineContent methods added in 4.4.3 in
https://github.com/scalameta/scalameta/pull/2197 , since such methods already existed in mdoc. I remvoed the lineContent one
and kept the lineCaret as a private method, because it's logic is a bit different than the scalameta one.