Make all older Scala versions work
Previously, we cross published 2.12.13 and for older versions we used 2.12.12 due to changes to compiler API. This however did not work for the sbt plugin, which had the coursier resolve a wrong version for a transitive dependency. Now, we are using a more base reporting API, which should work for both 2.12.12- and 2.12.13+ versions. I previously haven't noticed that this is possible.

The couriser issue is https://github.com/coursier/coursier/issues/1950