bugfix: [Scala 3] Reset context before each fail section
Previously, we were using the same reporter for each fail section, which was causing no more errors to be reported on the next section.

Now, we reset the context before compiling, which is creating a new reporter each time and that was already done for `compile`

I also tidied it a bit to only use the currently available context which will be fresh on `reset` always.

Fixes https://github.com/scalameta/mdoc/issues/663