MarkdownPart: move to separate mdoc-parser module
scalafmt has an own markdown parser, to be able to format any scala code
found in fenced blocks. In addition to duplicating functionality in that
parser, it also doesn't support indented code blocks.

Let's instead extract the parser portion of the mdoc logic into a module
so that it can be re-used. That module will have no dependencies.