bugfix: Remove additional whitespace from summary
The recent changes caused the spaces to no longer be removed properly due to the fact that we no longer operate on an actual value when printing one line summary.

I tried solving this multiple ways but:
- there is no sensible way to just copy any value
- using pprint.Tree as itnermediate representation doesn't work as those contain Iterator and are one use only

The easiest was to just replace newline with space and then deduplicate whitespace to save space on the online summary. This also makes it consistent across Scala 2 and 3